+++ docs/gen.sh
@@ -1,1 +1,2 @@
-make html -f
+#!/usr/bin/env bash
+make html